Minor fixes

+++ inward_supply/models.py
@@ -44,7 +44,7 @@ class Supplier(models.Model):
     email_id = models.EmailField()
     address = models.CharField(max_length=255)
     debit = models.FloatField(default=0.0)
-    total_sales = models.IntegerField(default=0)
+    total_sales = models.FloatField(default=0)
 
     def __str__(self):
         return f"{self.firm_name} ({self.person_name})"
